Fixing typos/grammar/wording
ln 6 and 36: Pluralized words for consistency. Note that the other headings that don't have 'a' or 'an' before it are all pluralized. eg closing bugs, issues, requests, etc. Questions or Problems also sounds slightly better.

ln 6: pluralized the links

ln 62: duplicate link

ln 127: themselves is gender neutral avoiding the him/herself so is a more succinct way of stating that they shouldn't merge their own pull request. I feel the themself should also be tacked on the end as "merging themselves the pull request" might sound slightly awkward. Similarly stating "or the admin" after "merge the branch" could also be awkward (and beg the question: or the admin what) and stating "the reviewer or the admin" flows better.

ln 143: "that much conventions" is grammatically incorrect, though not technically incorrect "we do not have yet" does not flow as well as "we do not yet have" I also added "When" before "in doubt" since on ln 170 you do use 'when in doubt' so I assume you just forgot "when"

ln 258: changed do to done -- tense agreement

ln 260: undefined reference: clarified what is needed, "installing pip" as what "if you need to" is referring to can be ambiguous

ln 287: changed 'user' to 'username' for consistency since in the next line 'username' is used

ln 376: I think this is what is meant, otherwise this is an odd sentence

ln 391: Windows

ln 403: ambiguous 'that'

ln 403, removed 'do'