Issue #1289. Delete user row when logging out.
We do this because GitHub gives us a unique access token each time
the user logs in. Otherwise, we end up with a lot of useless access
tokens in the user.db, which serve no purpose.

The only reason to keep the token in the db at all is to be able
to use it when performing actions against the GitHub API on the users
behalf.