Issue #1311 - Adds Issues DB for the project
Issues DB is (for now) contains only few data.
If you develop a new feature that would use this DB,
develop it in a way that makes possible to scrape the DB entirely.
Think about it as a cache that you can evict at any moment.
The project needs to be running even if the cache is empty.