Issue #1971 - Makes stricter the label construction
This doesn't solve everything, but it provides a way to mitigate some of the issues we met. It cut down a lot of possible values and variability. And the fallback being "fixme" if everything else fails, we can count on humans for guessing if possible the browser label.

This doesn't address yet the issue #2372 we will come to it later.