Issue #2479 - Makes report_issue uniform
Before report_issue was reporting either a JSON as dict, either a response object, this makes it uniform so it will only report JSON data as a dict. It simplifies testing and make it less… surprising. :)

If in the future we need the access to status, we can revisit this and makes it return response but uniformly.