Issue #2670 - Makes sure that we are dealing with a JSON object
Previously we failed if the value was for example 'null', which is a valid JSON value. So we make sure that the JSON is an array, because it is what we are expecting. And if it's not an array we just return a string.
Also if we get a ValueError we