Issue #2839 - Adds a new algorithm for computing fixtures checksum
The old checksum was based on a stringified dictionary and hence was variable depending on the way the dictionary would be serialized. We chose to create a checksum based on the URL full path.

The full URL would introduce a variability when computing the checksum.

When switching to python 3 we will need to make sure this is bytes oriented.