Issue #3146 - Adjusts the config to hardcode the unmoderated milestone
In the future, this should be probably in the `.env` or environment variable on the server.

Also it would be a possibility for us to clean up the code of initializing milestone. Aka instead of defining the milestone through http requests at the start of flask, we could just decide that the deploy script grab them if required and put them in env. Or even just write them down in server environment. Since we used them they never changed.