Issue #3140 - Adjusts the model for sending information
Basically we have a structured dictionary which contains all the info we need.
This will be sent to the repo for updating the public issue.

There's still a bit of data massaging to make it proper in `private_issue_moderation`

(Probably once this is all done, we would need to rebase all of this or create a new branch with a proper set of commits.)