issue #3360 - adds list of reported_with values
This makes a controled list of accepted values for the parameter reported_with which is added to the body of the bug report.

The reports which are out of this list will be either web (default) or unknown.

Ultimately we will move this list and a couple of others outside of the committed code but still loaded by the config file. So we can have more flexibility when it's time to add or retire a value without having to merge a pull request.