Issue #3380 - accepts list with 3 and 2 items tuples
Originally the list of tuples were 2 items long.
but the wizard is using in some cases 3 items long list (because of icons) for radio fields.

This leds to the duplication of the list of problem_choices with 2 items and 3 items. So instead of duplicating code, let's handle it in the function itself. There might be something more elegant to do. We can discover that later.