Issue #3326 - Adds a prototype of Issue model
The goal is to put in place the base of a tool that will help
us to create a frozen static version of an issue.

Issue(100) initializes an object for the issue 100
It will grab the parameter and create a ready to serialize object
into html or any other required format.

Probably in the future this model can probably be reused on
webcompat.com itself to represent issues.