issue #3326 - Separate Issue and ArchivedIssue
Probably Issue will move to a more general model elsewhere
that we can share in the context of the flask app.
So we will reduce the duplication effort.

This initializes from a dictionary instead of the idea
of fetching the data dynamically.

This also removes the check on the integer value of the issue number.
because we now initialize from data we would receive from a real JSON
converted as a dict.