issue #3326 - adds a save method for archived issues
The issues will be saved in /static/issues/issue-100.html for the issue 100.

it also makes sure that the directory exists before trying to save so we don't end up with a hard fail. If it doesn't exist it creates the directory.

The save method returns a Path object from pathlib
https://docs.python.org/3/library/pathlib.html

That we can eventually reuse for other things such as logging that the issue was saved.