issue #3326 - Adds skeleton for comments fetching
Comment is a dataclass model for each individual comments.
An issue should have a list of those when comments exist.

Once the issue has been instantiated, we could go fetch the comments
only if requested: `issue.fetch_comments(page=1)`
maybe we need `page=1` as a default parameters
and a special `page=all` for everything.

What is happening if the fetching fails totally? partially?
the comments_number could be handy for checking we fetched
all comments.

Should there be a model for one Comment as we do for issues?

On Issue probably we need something like.

`comments: List[Comment] = field(default_factory=create_comments_list)`

return a list of dictionary, each dictionary represents a comment.